Use ints in config struct instead of bools
fuse_opt_parse expects the struct fields to be ints.  When bools
are used, fuse_opt_parse writes beyond the field, clobbering
other fields or even writing beyond the boundary of the struct!

This error was not detected by the compiler because fuse_opt_parse does
not have a type-safe API :-(

Closes: #808779